additional PVP notification options
additional PVP notification options:
these only apply to the wilderness.
Configures whether or not friends trigger a notification
Configures whether or not clan members trigger a notification
Configures whether or not to trigger a beeping warning sound
Configures whether or not to draw an arrow on new people who have spawned
credit: LyzrdLite for the plugin, I just moved them to a better location.