Supersedes/Closes runelite#7353 Supersedes/Closes runelite#7415 Supersedes/Closes runelite#7493 Supersedes/Closes runelite#7401 Supersedes/Closes runelite#7151 Supersedes/Closes runelite#5383 streamline teleport item swaps
Move Mage of Zamorak teleport to Travel swap
It is a travel swap so there's no reason for it to have it's own option.

unify bank and exchange options
There's no reason for these to be separate options.

Swap Coal Bag Empty - Makes Empty the left click option when in a bank