Add a new function to the WorldArea api. Add npcstatus plugin.
The new function is for determining if it's possible to do a melee
attack between 2 actors of which the WorldArea are given. This checks
for melee distance AND for collision. That new function is used in
the npcstatus plugin.

The npcstatus plugin requires the attackrate and attackrange of an npc
as input, and returns a timer on the npc when they either are targetting
you or you are targetting them. With only the given input, the plugin is
able to determine when an npc will/can attack, when the npc can't be
flinched, when the npc can be flinched, and the duration of the flinch.

Timer info:
Orange: counter for when the npc is able to attack again. The timer is
caused by getting attacked by the npc.
Red: counter for when the npc will be flinchable again, assuming the
player won't get hit by the npc during this countdown.
Blue 0: There's no timer at this moment (just a 0), and when the player
causes a red/blue hitsplat on the npc, the flinch timer will start
(except when the npc initiates an attack before the hitsplat arrives).

Possible enhancements to be made in the future:
-Manually add some bosses to the plugin (like zammy, 6-tick, attackrange
1)
-Add the option to filter some npc's (for example in gwd, option to
filter out the minions)
-Reduce the amount of spaghetti in the code (this is basically my
first time using jave lol)
-Implement a learning algorithm to let the plugin determine the
attackdelay and attackrange of npc's (sounds difficult but possible to
do, but I won't put my time in it)
-Possibly improve the graphics of the timers

Final note: this is basically an improved version of the tick counter
plugin, because it's able to detect flinching mechanics and because it
does not depend on npc animation ID's.
It's also an improved version of
the flinch timer plugin because it's easier to know the attackrate and
(a good guess of) the attackrange than to know the exact timer duration,
unless you're one of the few people that know how flinching actually
works. It also works for non-melee flinching, unlike that plugin. For
more info about flinching: I wrote an article on
https://oldschool.runescape.wiki/w/Flinching