Renaming and Tests for Sound Effects in cache.
As it stands this extracts the sound effects to .wav format. Sound which like for instance 3286.wav could potentially be used for alerting a user when using certain plugins such as Timers or Notifications. Though something may need to be put in place to allow RuneLite to play the wav file instead of maybe trying to spoof the packet to play a sound effect. https://github.com/runelite/runelite/pull/3699