mixins: Fix posting of WidgetPressed event
Previously the WidgetPressed event was posted during a
FieldHook of the tempMenuAction field. During handling
of this event the MenuManager would call setTempMenuAction.

However, the game code then subsequently overrides the
actions done by the MenuManager by modifying the MenuEntry.

Replacing the FieldHook with a method hook fixes this issue
by slightly delaying the event.