client: Make overlays aware of mouse focus
This is helpful for plugins that wish to show overlays that perform
actions when they have mouse focus, such as offering left click actions.

While it's possible to track this within the overlay itself it is difficult
to do correctly when considering overlapping overlays and overlays being
removed while they have focus.